add AppStore packages

+++ install.sh
@@ -5,20 +5,25 @@ setup_homebrew() {
     ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
   fi
 
+  sudo chown -R $(whoami) /usr/local/lib/pkgconfig
   brew update
   brew bundle
 }
 
 setup_zsh() {
-  git clone --recursive https://github.com/sorin-ionescu/prezto.git "${ZDOTDIR:-$HOME}/.zprezto"
+  if [ ! -d "${ZDOTDIR:-$HOME}/.zprezto" ]; then
+    git clone --recursive https://github.com/sorin-ionescu/prezto.git "${ZDOTDIR:-$HOME}/.zprezto"
+  fi
 
   setopt EXTENDED_GLOB
   for rcfile in "${ZDOTDIR:-$HOME}"/.zprezto/runcoms/^README.md(.N); do
-    ln -s "$rcfile" "${ZDOTDIR:-$HOME}/.${rcfile:t}"
+    if [ ! -f "${ZDOTDIR:-$HOME}/.${rcfile:t}" ]; then
+      ln -s "$rcfile" "${ZDOTDIR:-$HOME}/.${rcfile:t}"
+    fi
   done
 
   chsh -s /bin/zsh
 }
 
-setup_homebrew
 setup_zsh
+setup_homebrew
